+++ src/main/java/com/github/choonchernlim/security/adfs/saml2/SpringSecurityAdfsSaml2Exception.java
@@ -1,0 +1,11 @@
+package com.github.choonchernlim.security.adfs.saml2;
+
+/**
+ * Module specific exception.
+ */
+public final class SpringSecurityAdfsSaml2Exception extends RuntimeException {
+
+    SpringSecurityAdfsSaml2Exception(final String message) {
+        super(message);
+    }
+}