+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 0.4.0 - 2016-06-01
+
+* If `samlConfigBean.samlUserDetailsService` is provided, then set `samlAuthenticationProvider.forcePrincipalAsString` to `false` so that `principal` represents the `userDetails` object.
+* Ability to mock security to bypass authentication against ADFS during rapid app development. To use this, `samlConfigBean.samlUserDetailsService` must be set.
+* Dependency, parent and plugins updates.
+
+```                  
+com.github.choonchernlim:build-reports ................ 0.2.4 -> 0.3.2
+com.google.guava:guava-testlib .......................... 18.0 -> 19.0
+junit:junit ............................................. 4.11 -> 4.12
+org.codehaus.groovy:groovy-all .............. 2.4.3 -> 2.4.6
+org.springframework.security:spring-security-config ...
+                                        4.0.3.RELEASE -> 4.1.0.RELEASE
+org.springframework.security:spring-security-core ...
+                                        4.0.3.RELEASE -> 4.1.0.RELEASE
+org.springframework.security:spring-security-web ...
+                                        4.0.3.RELEASE -> 4.1.0.RELEASE
+org.springframework.security.extensions:spring-security-saml2-core ...
+                                        1.0.1.RELEASE -> 1.0.2.RELEASE
+maven-compiler-plugin ................................... 3.3 -> 3.5.1
+```                       
+
 ## 0.3.3 - 2016-04-13                       
 * Inject Spring environment to get access to project properties file. ([#1](https://github.com/choonchernlim/spring-security-adfs-saml2/pull/1))